Reorganize @stimulus/core tests
