--- packages/@stimulus/core/test/bare_controller_test_case.ts
@@ -1,6 +1,0 @@
-import { Controller } from "../src/controller"
-import { ControllerTestCase } from "./controller_test_case"
-
-export class BareControllerTestCase extends ControllerTestCase<Controller> {
-  controllerConstructor = Controller
-}
