+++ packages/@stimulus/core/test/cases/application_test_case.ts
@@ -1,6 +1,6 @@
-import { Application } from "../src/application"
+import { Application } from "../../src/application"
 import { DOMTestCase } from "@stimulus/test"
-import { Schema, defaultSchema } from "../src/schema"
+import { Schema, defaultSchema } from "../../src/schema"
 
 class TestApplication extends Application {
   handleError(error: Error, message: string, detail: object) {
