+++ packages/@stimulus/core/test/cases/controller_test_case.ts
@@ -1,9 +1,10 @@
 import { ApplicationTestCase } from "./application_test_case"
-import { ControllerConstructor } from "../src/controller"
+import { Constructor } from "../../src/constructor"
+import { Controller } from "../../src/controller"
 
-export class ControllerTestCase<T> extends ApplicationTestCase {
+export class ControllerTests<T extends Controller> extends ApplicationTestCase {
   identifier: string | string[] = "test"
-  controllerConstructor!: ControllerConstructor
+  controllerConstructor!: Constructor<Controller>
   fixtureHTML = `<div data-controller="${this.identifiers.join(" ")}">`
 
   setupApplication() {
@@ -32,4 +33,12 @@ export class ControllerTestCase<T> extends ApplicationTestCase {
   get controllers(): T[] {
     return this.application.controllers as any as T[]
   }
+}
+
+export function ControllerTestCase(): Constructor<ControllerTests<Controller>>
+export function ControllerTestCase<T extends Controller>(constructor: Constructor<T>): Constructor<ControllerTests<T>>
+export function ControllerTestCase<T extends Controller>(constructor?: Constructor<T>): Constructor<ControllerTests<T>> {
+  return class extends ControllerTests<T> {
+    controllerConstructor = constructor || Controller as any
+  } as any
 }
