+++ packages/@stimulus/core/test/controllers/log_controller.ts
@@ -1,4 +1,4 @@
-import { Controller } from "../src/controller"
+import { Controller } from "../../src/controller"
 
 export type ActionLogEntry = {
   name: string
