+++ packages/@stimulus/core/test/index.ts
@@ -1,5 +1,5 @@
 import "@stimulus/polyfills"
 
-const context = require.context("./cases", true, /\.js$/)
+const context = require.context("./modules", true, /\.js$/)
 const modules = context.keys().map(key => context(key).default)
 modules.forEach(constructor => constructor.defineModule())
