+++ packages/@stimulus/core/test/modules/application_start_tests.ts
