+++ packages/@stimulus/core/test/modules/es6_tests.ts
@@ -1,5 +1,5 @@
-import { LogController } from "../log_controller"
-import { LogControllerTestCase } from "../log_controller_test_case"
+import { LogController } from "../controllers/log_controller"
+import { LogControllerTestCase } from "../cases/log_controller_test_case"
 
 export default class ES6Tests extends LogControllerTestCase {
   static shouldSkipTest(testName: string) {
