+++ packages/@stimulus/core/test/modules/lifecycle_tests.ts
@@ -1,4 +1,4 @@
-import { LogControllerTestCase } from "../log_controller_test_case"
+import { LogControllerTestCase } from "../cases/log_controller_test_case"
 
 export default class LifecycleTests extends LogControllerTestCase {
   controllerElement!: Element
