--- packages/@stimulus/core/test/target_controller_test_case.ts
@@ -1,6 +1,0 @@
-import { ControllerTestCase } from "./controller_test_case"
-import { TargetController } from "./target_controller"
-
-export class TargetControllerTestCase extends ControllerTestCase<TargetController> {
-  controllerConstructor = TargetController
-}
